DNP_tables.php
this will create the tables for your database. i will be uploading the table values to dropbox, however i seem to be having issues when importing the tables with a NULL space in the "has_LAB" and "has_TUTORIAL"  section.  That is i want the null space on certin entries (defult should work) but i am not getting it, it populates the area with a 0, also it seems to be messing up the values in these colums when imported.